+++ Source/WTF/wtf/PlatformHave.h
@@ -1788,3 +1788,7 @@
 #if PLATFORM(IOS) && !PLATFORM(IOS_SIMULATOR)
 #define HAVE_MOBILE_KEY_BAG 1
 #endif
+
+#if (PLATFORM(MAC) && __MAC_OS_X_VERSION_MIN_REQUIRED >= 140000)
+#define HAVE_CA_DISPLAY_LINK_MAC 1
+#endif
