+++ Source/WebKit/SourcesCocoa.txt
@@ -539,8 +539,9 @@ UIProcess/Launcher/cocoa/ProcessLauncherCocoa.mm
 
 UIProcess/mac/CorrectionPanel.mm
 UIProcess/mac/DisplayCaptureSessionManager.mm
-UIProcess/mac/DisplayLinkMac.cpp
+UIProcess/mac/DisplayLinkMac.mm
 UIProcess/mac/HighPerformanceGPUManager.mm
+UIProcess/mac/LegacyDisplayLinkMac.cpp
 UIProcess/mac/LegacySessionStateCoding.cpp
 UIProcess/mac/PageClientImplMac.mm
 UIProcess/mac/SecItemShimProxy.cpp
